Fix longitude coordinate ranges
The UM requires that longitude coordinates in island descriptions lie
in the range [2, NX+1], where NX is the number of grid cells in the
longitude direction on the tracer grid.  This commit makes Isla behave
correctly with respect to this constraint, in particular dealing
correctly with islands that span the meridian of longitude.